+++ api_yamdb/api_yamdb/settings.py
@@ -25,7 +25,6 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'django_filters',
-    'core.apps.CoreConfig',
     'reviews.apps.ReviewsConfig',
     'api.apps.ApiConfig',
 ]
